split classes to core
